requirements.txt
aiofiles==25.1.0
annotated-types==0.7.0
anyio==4.12.1
beautifulsoup4==4.14.3
certifi==2026.1.4
edgartools==5.16.1
filelock==3.24.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpxthrottlecache==0.3.0
humanize==4.15.0
idna==3.11
Jinja2==3.1.6
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
nest-asyncio==1.6.0
numpy==2.4.2
orjson==3.11.7
pandas==3.0.1
pyarrow==23.0.1
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyrate-limiter==4.0.2
python-dateutil==2.9.0.post0
rank-bm25==0.2.2
RapidFuzz==3.14.3
rich==14.3.2
six==1.17.0
soupsieve==2.8.3
stamina==25.2.0
tabulate==0.9.0
tenacity==9.1.4
textdistance==4.6.3
tqdm==4.67.3
truststore==0.10.4
typing-inspection==0.4.2
typing_extensions==4.15.0
Unidecode==1.4.0
nltk==3.9.1
openpyxl==3.1.5
scikit-learn==1.6.1
matplotlib==3.10.0
seaborn==0.13.2